Import Mesa 20.1.6

+++ lib/mesa/src/gallium/auxiliary/driver_rbug/rbug_core.c
@@ -27,7 +27,7 @@
 
 
 #include "os/os_thread.h"
-#include "util/u_format.h"
+#include "util/format/u_format.h"
 #include "util/u_string.h"
 #include "util/u_inlines.h"
 #include "util/u_memory.h"
@@ -44,8 +44,8 @@
 
 #include <errno.h>
 
-#define U642VOID(x) ((void *)(unsigned long)(x))
-#define VOID2U64(x) ((uint64_t)(unsigned long)(x))
+#define U642VOID(x) ((void *)(uintptr_t)(x))
+#define VOID2U64(x) ((uint64_t)(uintptr_t)(x))
 
 #define container_of(ptr, type, field) \
    (type*)((char*)ptr - offsetof(type, field))
